Fix Admin Bar menu options (#76)
This Pull Request introduces several improvements to how options are
managed within the `CheckboxGroup`, particularly in relation to the
Admin Bar menu. With this update, any newly added or removed of the menu
in the Admin Bar will be shown correctly on the list. It will also now
correctly add a menu item to the Admin Bar and set it to "enabled" by
default, rather than "disabled".